Start reordering the ispyb service
